requirements.txt
# Core ETL
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
openpyxl==3.1.5
unidecode==1.3.8

# Data Warehouse
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# Validación
# GE 1.x es requerido en Python 3.13 (0.18.x requiere numpy<2.0, incompatible con Py3.13)
# En Colab con Python 3.12 puede usarse: great-expectations==0.18.21
great-expectations>=1.15.0

# Visualización (EDA)
matplotlib==3.9.4
seaborn==0.13.2

# Variables de entorno
python-dotenv==1.0.1

# Airflow (instalar en entorno Linux/WSL/Docker)
# apache-airflow==2.10.4

# Jupyter (para notebooks)
jupyter==1.1.1
ipykernel==6.29.5

# Profiling (EDA)
ydata-profiling==4.10.0
